build/dir: rename button directory and update imports
